fix import and mocked querystring_options in patterns tests
