+++ plone/app/z3cform/tests/test_widgets.py
@@ -532,7 +532,7 @@ class SelectWidgetTests(unittest.TestCase):
         base.unregisterAdapter(term.CollectionTermsSource)
 
     def test_widget(self):
-        from plone.app.z3cform.widget import SelectWidget
+        from plone.app.z3cform.widgets.select import SelectWidget
 
         widget = SelectWidget(self.request)
         widget.id = "test-widget"
@@ -630,7 +630,7 @@ class SelectWidgetTests(unittest.TestCase):
         )
 
     def test_widget_list_orderable(self):
-        from plone.app.z3cform.widget import SelectWidget
+        from plone.app.z3cform.widgets.select import SelectWidget
 
         widget = SelectWidget(self.request)
         widget.id = "test-widget"
@@ -657,7 +657,7 @@ class SelectWidgetTests(unittest.TestCase):
         )
 
     def test_widget_tuple_orderable(self):
-        from plone.app.z3cform.widget import SelectWidget
+        from plone.app.z3cform.widgets.select import SelectWidget
 
         widget = SelectWidget(self.request)
         widget.id = "test-widget"
@@ -683,7 +683,7 @@ class SelectWidgetTests(unittest.TestCase):
         )
 
     def test_widget_set_not_orderable(self):
-        from plone.app.z3cform.widget import SelectWidget
+        from plone.app.z3cform.widgets.select import SelectWidget
 
         widget = SelectWidget(self.request)
         widget.id = "test-widget"
@@ -710,7 +710,7 @@ class SelectWidgetTests(unittest.TestCase):
         )
 
     def test_widget_extract(self):
-        from plone.app.z3cform.widget import SelectWidget
+        from plone.app.z3cform.widgets.select import SelectWidget
 
         widget = SelectWidget(self.request)
         widget.field = Choice(
@@ -726,7 +726,7 @@ class SelectWidgetTests(unittest.TestCase):
 
     def test_data_converter_list(self):
         from plone.app.z3cform.converters import SelectWidgetConverter
-        from plone.app.z3cform.widget import SelectWidget
+        from plone.app.z3cform.widgets.select import SelectWidget
 
         field = List(
             __name__="listfield",
@@ -773,7 +773,7 @@ class SelectWidgetTests(unittest.TestCase):
 
     def test_data_converter_tuple(self):
         from plone.app.z3cform.converters import SelectWidgetConverter
-        from plone.app.z3cform.widget import SelectWidget
+        from plone.app.z3cform.widgets.select import SelectWidget
 
         field = Tuple(
             __name__="tuplefield",
@@ -809,7 +809,7 @@ class SelectWidgetTests(unittest.TestCase):
 
     def test_data_converter_handles_empty_value(self):
         from plone.app.z3cform.converters import SelectWidgetConverter
-        from plone.app.z3cform.widget import SelectWidget
+        from plone.app.z3cform.widgets.select import SelectWidget
 
         field = Tuple(
             __name__="tuplefield",
@@ -829,7 +829,7 @@ class SelectWidgetTests(unittest.TestCase):
         """
         If the widget vocabulary is a mapping <optgroup>'s are rendered.
         """
-        from plone.app.z3cform.widget import SelectWidget
+        from plone.app.z3cform.widgets.select import SelectWidget
         from z3c.form import term
 
         widget = SelectWidget(self.request)
@@ -885,7 +885,7 @@ class AjaxSelectWidgetTests(unittest.TestCase):
         provideUtility(example_vocabulary_factory, name="example")
 
     def test_widget(self):
-        from plone.app.z3cform.widget import AjaxSelectWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectWidget
 
         widget = AjaxSelectWidget(self.request)
         widget.update()
@@ -932,7 +932,7 @@ class AjaxSelectWidgetTests(unittest.TestCase):
         )
 
     def test_widget_list_orderable(self):
-        from plone.app.z3cform.widget import AjaxSelectWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectWidget
 
         widget = AjaxSelectWidget(self.request)
         widget.field = List(__name__="selectfield")
@@ -947,7 +947,7 @@ class AjaxSelectWidgetTests(unittest.TestCase):
         )
 
     def test_widget_tuple_orderable(self):
-        from plone.app.z3cform.widget import AjaxSelectWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectWidget
 
         widget = AjaxSelectWidget(self.request)
         widget.field = Tuple(__name__="selectfield")
@@ -962,7 +962,7 @@ class AjaxSelectWidgetTests(unittest.TestCase):
         )
 
     def test_widget_set_not_orderable(self):
-        from plone.app.z3cform.widget import AjaxSelectWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectWidget
 
         widget = AjaxSelectWidget(self.request)
         # A set is not orderable
@@ -978,7 +978,7 @@ class AjaxSelectWidgetTests(unittest.TestCase):
         )
 
     def test_widget_choice(self):
-        from plone.app.z3cform.widget import AjaxSelectWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectWidget
         from zope.schema.interfaces import ISource
 
         widget = AjaxSelectWidget(self.request)
@@ -1002,7 +1002,7 @@ class AjaxSelectWidgetTests(unittest.TestCase):
         )
 
     def test_widget_addform_url_on_addform(self):
-        from plone.app.z3cform.widget import AjaxSelectWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectWidget
 
         widget = AjaxSelectWidget(self.request)
         form = Mock(parentForm=None)
@@ -1037,7 +1037,7 @@ class AjaxSelectWidgetTests(unittest.TestCase):
 
     def test_data_converter_list(self):
         from plone.app.z3cform.converters import AjaxSelectWidgetConverter
-        from plone.app.z3cform.widget import AjaxSelectWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectWidget
 
         field = List(__name__="listfield", value_type=TextLine())
         widget = AjaxSelectWidget(self.request)
@@ -1066,7 +1066,7 @@ class AjaxSelectWidgetTests(unittest.TestCase):
 
     def test_data_converter_collection_with_vocabulary(self):
         from plone.app.z3cform.converters import AjaxSelectWidgetConverter
-        from plone.app.z3cform.widget import AjaxSelectWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectWidget
 
         field = Tuple(
             __name__="listfield",
@@ -1100,7 +1100,7 @@ class AjaxSelectWidgetTests(unittest.TestCase):
 
     def test_data_converter_tuple(self):
         from plone.app.z3cform.converters import AjaxSelectWidgetConverter
-        from plone.app.z3cform.widget import AjaxSelectWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectWidget
 
         field = Tuple(__name__="tuplefield", value_type=TextLine())
         widget = AjaxSelectWidget(self.request)
@@ -1128,8 +1128,8 @@ class AjaxSelectWidgetTests(unittest.TestCase):
         )
 
     def test_fieldwidget(self):
-        from plone.app.z3cform.widget import AjaxSelectFieldWidget
-        from plone.app.z3cform.widget import AjaxSelectWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectFieldWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectWidget
 
         field = Mock(__name__="field", title="", required=True)
         request = Mock()
@@ -1139,8 +1139,8 @@ class AjaxSelectWidgetTests(unittest.TestCase):
         self.assertIs(widget.request, request)
 
     def test_fieldwidget_sequence(self):
-        from plone.app.z3cform.widget import AjaxSelectFieldWidget
-        from plone.app.z3cform.widget import AjaxSelectWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectFieldWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectWidget
 
         field = Mock(__name__="field", title="", required=True)
         vocabulary = Mock()
@@ -1158,7 +1158,7 @@ class AjaxSelectWidgetIntegrationTests(unittest.TestCase):
         self.request = TestRequest(environ={"HTTP_ACCEPT_LANGUAGE": "en"})
 
     def test_keywords_can_add(self):
-        from plone.app.z3cform.widget import AjaxSelectWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectWidget
 
         portal = self.layer["portal"]
         setRoles(portal, TEST_USER_ID, ["Manager"])
@@ -1171,7 +1171,7 @@ class AjaxSelectWidgetIntegrationTests(unittest.TestCase):
         )
 
     def test_keywords_cannot_add(self):
-        from plone.app.z3cform.widget import AjaxSelectWidget
+        from plone.app.z3cform.widgets.select import AjaxSelectWidget
 
         portal = self.layer["portal"]
         widget = AjaxSelectWidget(self.request)
@@ -1181,6 +1181,17 @@ class AjaxSelectWidgetIntegrationTests(unittest.TestCase):
             widget._base_args()["pattern_options"]["allowNewItems"],
             "false",
         )
+
+
+def mock_querystring_options(context, querystring_view):
+    return {
+        'indexOptionsUrl': '/{}'.format(querystring_view),
+        'previewURL': '/@@querybuilder_html_results',
+        'previewCountURL': '/@@querybuildernumberofresults',
+        'patternDateOptions': None,
+        'patternAjaxSelectOptions': {'separator': ';'},
+        'patternRelateditemsOptions': None,
+    }
 
 
 class QueryStringWidgetTests(unittest.TestCase):
@@ -1202,12 +1213,8 @@ class QueryStringWidgetTests(unittest.TestCase):
         self.assertEqual(converter.toFieldValue("[]"), None)
 
     @mock.patch(
-        "plone.app.z3cform.widgets.datetime.get_date_options",
-        new=lambda *args, **kwargs: None,
-    )
-    @mock.patch(
-        "plone.app.z3cform.widgets.relateditems.get_relateditems_options",
-        new=lambda *args, **kwargs: None,
+        "plone.app.z3cform.widgets.querystring.get_querystring_options",
+        new=mock_querystring_options,
     )
     def test_widget(self):
         from plone.app.z3cform.widgets.querystring import QueryStringWidget
@@ -1249,7 +1256,7 @@ class RelatedItemsWidgetIntegrationTests(unittest.TestCase):
         return set(expected.items()).issubset(set(actual.items()))
 
     def test_related_items_widget(self):
-        from plone.app.z3cform.widget import RelatedItemsWidget
+        from plone.app.z3cform.widgets.relateditems import RelatedItemsWidget
 
         EXPECTED_ROOT_PATH = "/plone"
         EXPECTED_ROOT_URL = "http://nohost/plone"
@@ -1281,7 +1288,7 @@ class RelatedItemsWidgetIntegrationTests(unittest.TestCase):
 
     def test_related_items_widget_nav_root(self):
         from plone.app.layout.navigation.interfaces import INavigationRoot
-        from plone.app.z3cform.widget import RelatedItemsWidget
+        from plone.app.z3cform.widgets.relateditems import RelatedItemsWidget
 
         EXPECTED_ROOT_PATH = "/plone"
         EXPECTED_ROOT_URL = "http://nohost/plone"
@@ -1434,7 +1441,7 @@ class RelatedItemsWidgetTests(unittest.TestCase):
     def test_multiple_selection(self):
         """The pattern_options key maximumSelectionSize shouldn't be
         set when the field allows multiple selections"""
-        from plone.app.z3cform.widget import RelatedItemsFieldWidget
+        from plone.app.z3cform.widgets.relateditems import RelatedItemsFieldWidget
         from zope.schema.interfaces import ISource
         from zope.schema.vocabulary import VocabularyRegistry
 
@@ -1561,8 +1568,8 @@ class RelatedItemsWidgetTests(unittest.TestCase):
             )
 
     def test_fieldwidget(self):
-        from plone.app.z3cform.widget import RelatedItemsFieldWidget
-        from plone.app.z3cform.widget import RelatedItemsWidget
+        from plone.app.z3cform.widgets.relateditems import RelatedItemsFieldWidget
+        from plone.app.z3cform.widgets.relateditems import RelatedItemsWidget
 
         field = Mock(__name__="field", title="", required=True)
         vocabulary = Mock()
@@ -1584,7 +1591,7 @@ def add_mock_fti(portal):
 
 
 def _custom_field_widget(field, request):
-    from plone.app.z3cform.widget import AjaxSelectWidget
+    from plone.app.z3cform.widgets.select import AjaxSelectWidget
 
     widget = FieldWidget(field, AjaxSelectWidget(request))
     widget.vocabulary = "plone.app.vocabularies.PortalTypes"
@@ -1610,7 +1617,7 @@ class RichTextWidgetTests(unittest.TestCase):
         self.field = IWithText["text"]
 
     def test_widget_params(self):
-        from plone.app.z3cform.widget import RichTextWidget
+        from plone.app.z3cform.widgets.richtext import RichTextWidget
 
         widget = FieldWidget(self.field, RichTextWidget(self.request))
         # set the context so we can get tinymce settings
@@ -1632,7 +1639,7 @@ class RichTextWidgetTests(unittest.TestCase):
         )
 
     def test_widget_params_different_contexts(self):
-        from plone.app.z3cform.widget import RichTextWidget
+        from plone.app.z3cform.widgets.richtext import RichTextWidget
 
         setRoles(self.portal, TEST_USER_ID, ["Contributor"])
 
@@ -1683,7 +1690,7 @@ class RichTextWidgetTests(unittest.TestCase):
 
     def test_widget_values(self):
         from plone.app.textfield.value import RichTextValue
-        from plone.app.z3cform.widget import RichTextWidget
+        from plone.app.z3cform.widgets.richtext import RichTextWidget
 
         widget = FieldWidget(self.field, RichTextWidget(self.request))
         # set the context so we can get tinymce settings
@@ -1710,7 +1717,7 @@ class RichTextWidgetTests(unittest.TestCase):
         if IMarkupSchema:
             # if not, don't run this test
             self._set_mimetypes(allowed=("text/html",))
-            from plone.app.z3cform.widget import RichTextWidget
+            from plone.app.z3cform.widgets.richtext import RichTextWidget
 
             widget = FieldWidget(self.field, RichTextWidget(self.request))
             # set the context so we can get tinymce settings
@@ -1729,7 +1736,7 @@ class RichTextWidgetTests(unittest.TestCase):
         if IMarkupSchema:
             # if not, don't run this test
             self._set_mimetypes(allowed=("text/html", "text/plain"))
-            from plone.app.z3cform.widget import RichTextWidget
+            from plone.app.z3cform.widgets.richtext import RichTextWidget
 
             widget = FieldWidget(self.field, RichTextWidget(self.request))
             # set the context so we can get tinymce settings
@@ -1753,7 +1760,7 @@ class RichTextWidgetTests(unittest.TestCase):
             # if not, don't run this test
             self._set_mimetypes(allowed=("text/html", "text/plain"))
             from plone.app.textfield.value import RichTextValue
-            from plone.app.z3cform.widget import RichTextWidget
+            from plone.app.z3cform.widgets.richtext import RichTextWidget
 
             widget = FieldWidget(self.field, RichTextWidget(self.request))
             # set the context so we can get tinymce settings
@@ -1780,7 +1787,7 @@ class RichTextWidgetTests(unittest.TestCase):
             proxy = registry.forInterface(IEditingSchema, check=False, prefix="plone")
             proxy.available_editors = ["dummy", "TinyMCE"]
             proxy.default_editor = "dummy"
-            from plone.app.z3cform.widget import RichTextWidget
+            from plone.app.z3cform.widgets.richtext import RichTextWidget
 
             widget = FieldWidget(self.field, RichTextWidget(self.request))
             widget.context = self.portal
@@ -1788,7 +1795,7 @@ class RichTextWidgetTests(unittest.TestCase):
             self.assertTrue("<p>dummy</p>" in rendered)
 
             proxy.default_editor = "TinyMCE"
-            from plone.app.z3cform.widget import RichTextWidget
+            from plone.app.z3cform.widgets.richtext import RichTextWidget
 
             widget = FieldWidget(self.field, RichTextWidget(self.request))
             widget.context = self.portal
@@ -1806,7 +1813,7 @@ class LinkWidgetIntegrationTests(unittest.TestCase):
         setRoles(self.portal, TEST_USER_ID, ["Manager"])
 
     def test_link_widget__pattern_options(self):
-        from plone.app.z3cform.widget import LinkWidget
+        from plone.app.z3cform.widgets.link import LinkWidget
 
         widget = LinkWidget(self.request)
 
@@ -1818,7 +1825,7 @@ class LinkWidgetIntegrationTests(unittest.TestCase):
         self.assertEqual(pattern_data["maximumSelectionSize"], 1)
 
     def test_link_widget__extract_internal(self):
-        from plone.app.z3cform.widget import LinkWidget
+        from plone.app.z3cform.widgets.link import LinkWidget
 
         widget = LinkWidget(self.request)
         widget.context = self.portal
@@ -1832,7 +1839,7 @@ class LinkWidgetIntegrationTests(unittest.TestCase):
         )
 
     def test_link_widget__extract_external(self):
-        from plone.app.z3cform.widget import LinkWidget
+        from plone.app.z3cform.widgets.link import LinkWidget
 
         widget = LinkWidget(self.request)
         widget.context = self.portal
@@ -1846,7 +1853,7 @@ class LinkWidgetIntegrationTests(unittest.TestCase):
         )
 
     def test_link_widget__extract_email(self):
-        from plone.app.z3cform.widget import LinkWidget
+        from plone.app.z3cform.widgets.link import LinkWidget
 
         widget = LinkWidget(self.request)
         widget.context = self.portal
@@ -1860,7 +1867,7 @@ class LinkWidgetIntegrationTests(unittest.TestCase):
         )
 
     def test_link_widget__extract_email_including_mailto(self):
-        from plone.app.z3cform.widget import LinkWidget
+        from plone.app.z3cform.widgets.link import LinkWidget
 
         widget = LinkWidget(self.request)
         widget.context = self.portal
@@ -1875,7 +1882,7 @@ class LinkWidgetIntegrationTests(unittest.TestCase):
 
     def test_link_widget__data_converter(self):
         from plone.app.z3cform.converters import LinkWidgetDataConverter
-        from plone.app.z3cform.widget import LinkWidget
+        from plone.app.z3cform.widgets.link import LinkWidget
 
         field = TextLine(__name__="linkfield")
         widget = LinkWidget(self.request)
